requirements.txt
# Zerodha Kite API
kiteconnect>=4.0.0

# Data analysis and manipulation
pandas>=1.5.0
numpy>=1.23.0

# Visualization and Dashboard
plotly>=5.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0

# Utilities
python-dotenv>=0.21.0

# Machine Learning and AI
scikit-learn>=1.3.0

# Optional: For advanced technical analysis
# ta-lib>=0.4.24
# pandas-ta>=0.3.14

# Optional: For advanced NLP sentiment analysis  
# textblob>=0.17.1
# vaderSentiment>=3.3.2
